requirements_basic.txt
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==11.1.0
numpy==2.2.3
requests==2.32.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
opencv-python==4.9.0.80 